Component's actions can now get the initiator's name from the context
ConsoleComponent is not part of the browser core anymore since it's just
for the showcase.

Xissle.parseDom now takes an object of custom components to parse from the DOM.

Actions' context now contains a `from` property and `internals` property
which is a function that returns the component's name, storage and the
names of the groups it's subscribed to. Before they used to be a part of
each context. `currentGroup` is renamed to `group`.